:art: rename jira create_ticket command
